+++ src/main/java/io/github/torand/fastersql/dialect/H2Dialect.java
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static io.github.torand.fastersql.dialect.Capability.CONCAT_OPERATOR;
+import static io.github.torand.fastersql.dialect.Capability.CURRENT_TIME;
 import static io.github.torand.fastersql.dialect.Capability.LIMIT_OFFSET;
 import static io.github.torand.fastersql.util.lang.StringHelper.generate;
 
@@ -32,7 +33,7 @@ import static io.github.torand.fastersql.util.lang.StringHelper.generate;
  * <a href="https://www.h2database.com/html/grammar.html" />
  */
 public class H2Dialect implements Dialect {
-    private static final EnumSet<Capability> SUPPORTED_CAPS = EnumSet.of(LIMIT_OFFSET, CONCAT_OPERATOR);
+    private static final EnumSet<Capability> SUPPORTED_CAPS = EnumSet.of(LIMIT_OFFSET, CONCAT_OPERATOR, CURRENT_TIME);
 
     @Override
     public String getProductName() {